+++ learning/java/language-essentials-slice/src/main/java/dev/frankamente/eos/learning/essentials/PracticeStatus.java
@@ -1,0 +1,26 @@
+package dev.frankamente.eos.learning.essentials;
+
+public enum PracticeStatus {
+  TODO,
+  READY,
+  IN_PROGRESS,
+  DONE,
+  BLOCKED;
+
+  public boolean isOpen() {
+    return switch (this) {
+      case TODO, READY, IN_PROGRESS -> true;
+      case DONE, BLOCKED -> false;
+    };
+  }
+
+  public String nextAction() {
+    return switch (this) {
+      case TODO -> "prepare";
+      case READY -> "start";
+      case IN_PROGRESS -> "finish";
+      case DONE -> "reflect";
+      case BLOCKED -> "unblock";
+    };
+  }
+}